@@ -18,12 +18,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   image_picker: ^0.7.4
   image_cropper: ^1.4.0
   camera: ^0.8.1
+  shared_preferences: ^2.0.5
+  path_provider: ^2.0.1
   flutter_launcher_icons: "^0.8.0"
   flutter:
     sdk: flutter
@@ -54,6 +56,7 @@ flutter:
   # To add assets to your application, add an assets section, like this:
   assets:
      - images/logo.PNG
+     - images/mina.jpg
  
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
